feat!: add RuntimeObject wsh no-restricted-globals
RuntimeObject is a JScript-specific global function which returns
variables and functions defined in the global scope (possibly filtered
based on its argument).

Instead of using this function, code should use `this`, which refers to
the global object in scopes where it is not provided by the caller and
is fully portable.  The StackOverflow post linked in the message
explains this, along with how to get the global scope in strict mode.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>