test: Ensure peerDependencies match devDependencies
Until https://github.com/eslint/eslint/issues/3458 is resolved, this
package must declare peerDependencies on configs and plugins used.  For
npm@^3 it must also declare them as devDependencies for them to be
installed during development/testing (see
https://github.com/npm/npm/issues/6565#issuecomment-74971689).

Therefore, all peerDependencies should be included in devDependencies
with the same version range so that the dev/test environment matches how
it will be used.  Test that this is the case using peer-version-check,
to catch mismatched versions (which I have committed multiple times in
the past).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>