test: exclude greenkeeper from test-peer-versions
Greenkeeper does not currently update peerDependencies (see
greenkeeperio/greenkeeper#151), which causes peer-version-check to fail
the tests every time Greenkeeper updates devDependencies which is also
peerDependencies.  (Which happens whenever Greenkeeper tests a new
version.)

Adding peerDependencies to greenkeeper.ignore is not a solution, since
we want to be informed when peerDependencies are out of date or cause
tests to fail.

Instead, ignore the peer-version-check exit code when run from Travis CI
for a greenkeeper branch.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>