feat: disable unicorn/no-for-loop
Due to conflicting with the restricted-syntax rule from
eslint-config-airbnb-base which prevents iterators/generators due to
regenerator polyfill overhead.

Reconsidering prohibitions on async/iterators/generators when the code
is only expected to be used in environments which don't require
transpilation/polyfills has been on the TODO list for a while.  This is
yet another example.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>