fix: rm arguments.callee no-restricted-properties
arguments.callee was disallowed by both the built-in `no-caller` rule
(which disallows both arguments.callee and arguments.caller) and by the
no-restricted-properties rule with the airbnb configuration.  This does
not add value and makes it more problematic to disable the rule, which
requires disabling both rules and has the usually-unintended effect of
disabling all no-restricted-properties matches in the disabled scope.

Remove arguments.callee from no-restricted-properties to avoid this.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>