feat: Allow/Prefer for...of loops
ForOfStatement is disallowed by airbnb-base both to encourage use of
immutable/functional Array methods (.every/filter/find/map/some) and to
avoid requiring regenerator-runtime in transpiled code.  Since I rarely
use transpilation or target engines which don't support generators,
regenerator-runtime is not a concern.  Although I agree with the
preference for immutable/functional methods, there are enough situations
where it is useful that the cost of forbidding for...of outweighs the
benefits.  Allow it, and prefer it to indexed for loops.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>