feat: Disable no-loop-func for all but WSH
Without var (which is disallowed by the no-var rule) the capturing
behavior is not confusing.  Since such declarations are allowed by ES6
supported on all engines I use, and stylistically preferred, disable
this rule.

Note: WSH is an exception, since it does not support const/let and
retains the potentially confusing behavior.  Therefore, continue to
enforce no-loop-func in WSH environment.

Note2:  IE 11 has an ugly "gotcha" that variables declared in the loop
initializer are treated as being outside of the block scope, which can
be a surprise for such functions.  An easy workaround is to copy the
value to a variable declared in the loop.  If this becomes an issue, I
may re-enable this rule for IE.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>