rework require() for peerDependencies
Loading configs/plugins required by this module is common.  ESLint
requires these to be peerDependencies, since they must be resolvable
from the project using this module.  That makes them difficult to
require from this module (e.g. for partially overriding configuration,
or making configuration dependent on logic in those modules).

Refactor the logic for loading such modules into a separate
eslint-require module so it can be reused as necessary.  Simplify the
current logic to use require.main.require and document ideas for
alternative logic if this proves to be insufficient in the future.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>