fix: replace eslint-require with require-peer
The assumption that require.main is the eslint module was a bad
assumption, since other tools (like depcheck) may load the ESLint
configuration.

After additional consideration, use a similar approach to
parent-require, without using it directly (since it loads from its own
module parent, which is not necessarily this module).  Use `require()`
as fall-back, in case the dependency can not be loaded from known module
ancestors (e.g. due to using the ESM loader, which does not set
module.parent).

Hope that this fix is sufficient until the new ESLint configuration
format is implemented and we can put the peerDependency madness behind
us.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>