feat: n/no-unsupported-features/node-builtins ignore fetch
[Global fetch](https://nodejs.org/api/globals.html#fetch) has been
supported without the `--experimental-fetch` CLI flag since Node.js
18.0.0.  Although it was considered experimental until 21.0.0, there
are no functional changes that I am aware of.  I see no risk of using it
on projects targeting Node.js >=18.

FIXME: fetch should not be ignored for Node.js <18, but there does not
appear to be a way to configure this.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>