Vendor eslint-config-airbnb-base@15.0.0
eslint-config-airbnb-base@15.0.0 specifies a peerDependency of
`eslint@^7.32.0 || ^8.2.0`. This causes warnings (or errors with
`--strict-peer-deps`) if eslint@^9 is used instead.

Until https://github.com/airbnb/javascript/issues/2961 is resolved,
use a vendored copy of eslint-config-airbnb-base to avoid this.

Note: https://github.com/NishargShah/eslint-config-airbnb-extended may
be a viable option, although it is currently incompatible with
@stylistic/eslint-plugin >= 4 due to lack of ESM support and it's
unclear whether it will receive sufficient adoption, support, and
community interest to keep it viable.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>